+++ packages/camera/camera/android/src/main/java/io/flutter/plugins/camera/Camera.java
@@ -230,7 +230,7 @@ public class Camera {
     final File file;
     try {
       file = File.createTempFile("CAP", ".jpg", outputDir);
-    } catch (IOException e) {
+    } catch (IOException | SecurityException e) {
       pictureCaptureRequest.error("cannotCreateFile", e.getMessage(), null);
       return;
     }
@@ -436,7 +436,7 @@ public class Camera {
     final File outputDir = applicationContext.getCacheDir();
     try {
       videoRecordingFile = File.createTempFile("REC", ".mp4", outputDir);
-    } catch (IOException e) {
+    } catch (IOException | SecurityException e) {
       result.error("cannotCreateFile", e.getMessage(), null);
       return;
     }