+++ packages/camera/camera/lib/camera.dart
@@ -6,4 +6,10 @@ export 'src/camera_controller.dart';
 export 'src/camera_image.dart';
 export 'src/camera_preview.dart';
 
-export 'package:camera_platform_interface/camera_platform_interface.dart';
+export 'package:camera_platform_interface/camera_platform_interface.dart'
+    show
+        CameraDescription,
+        CameraException,
+        CameraLensDirection,
+        ResolutionPreset,
+        XFile;