+++ packages/camera/camera/test/camera_test.dart
@@ -6,13 +6,12 @@ import 'dart:async';
 import 'dart:ui';
 
 import 'package:camera/camera.dart';
+import 'package:camera_platform_interface/camera_platform_interface.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
-
-import 'utils/method_channel_mock.dart';
 
 get mockAvailableCameras => [
       CameraDescription(
@@ -41,6 +40,7 @@ bool mockPlatformException = false;
 
 void main() {
   WidgetsFlutterBinding.ensureInitialized();
+
   group('camera', () {
     test('debugCheckIsDisposed should not throw assertion error when disposed',
         () {
@@ -245,159 +245,6 @@ void main() {
             'bar',
           )));
       mockPlatformException = false;
-    });
-
-    test('startImageStream() throws $CameraException when uninitialized', () {
-      CameraController cameraController = CameraController(
-          CameraDescription(
-              name: 'cam',
-              lensDirection: CameraLensDirection.back,
-              sensorOrientation: 90),
-          ResolutionPreset.max);
-      expect(
-          cameraController.startImageStream((image) => null),
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'Uninitialized CameraController.',
-            'startImageStream was called on uninitialized CameraController.',
-          )));
-    });
-
-    test('startImageStream() throws $CameraException when recording videos',
-        () async {
-      CameraController cameraController = CameraController(
-          CameraDescription(
-              name: 'cam',
-              lensDirection: CameraLensDirection.back,
-              sensorOrientation: 90),
-          ResolutionPreset.max);
-      await cameraController.initialize();
-
-      cameraController.value =
-          cameraController.value.copyWith(isRecordingVideo: true);
-      expect(
-          cameraController.startImageStream((image) => null),
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'A video recording is already started.',
-            'startImageStream was called while a video is being recorded.',
-          )));
-    });
-    test(
-        'startImageStream() throws $CameraException when already streaming images',
-        () async {
-      CameraController cameraController = CameraController(
-          CameraDescription(
-              name: 'cam',
-              lensDirection: CameraLensDirection.back,
-              sensorOrientation: 90),
-          ResolutionPreset.max);
-      await cameraController.initialize();
-
-      cameraController.value =
-          cameraController.value.copyWith(isStreamingImages: true);
-      expect(
-          cameraController.startImageStream((image) => null),
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'A camera has started streaming images.',
-            'startImageStream was called while a camera was streaming images.',
-          )));
-    });
-
-    test('startImageStream() calls CameraPlatform', () async {
-      MethodChannelMock methodChannelMock = MethodChannelMock(
-          channelName: 'plugins.flutter.io/camera',
-          methods: {'startImageStream': {}});
-      CameraController cameraController = CameraController(
-          CameraDescription(
-              name: 'cam',
-              lensDirection: CameraLensDirection.back,
-              sensorOrientation: 90),
-          ResolutionPreset.max);
-      await cameraController.initialize();
-
-      await cameraController.startImageStream((image) => null);
-
-      expect(methodChannelMock.log,
-          <Matcher>[isMethodCall('startImageStream', arguments: null)]);
-    });
-
-    test('stopImageStream() throws $CameraException when uninitialized', () {
-      CameraController cameraController = CameraController(
-          CameraDescription(
-              name: 'cam',
-              lensDirection: CameraLensDirection.back,
-              sensorOrientation: 90),
-          ResolutionPreset.max);
-
-      expect(
-          cameraController.stopImageStream(),
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'Uninitialized CameraController.',
-            'stopImageStream was called on uninitialized CameraController.',
-          )));
-    });
-
-    test('stopImageStream() throws $CameraException when recording videos',
-        () async {
-      CameraController cameraController = CameraController(
-          CameraDescription(
-              name: 'cam',
-              lensDirection: CameraLensDirection.back,
-              sensorOrientation: 90),
-          ResolutionPreset.max);
-      await cameraController.initialize();
-
-      await cameraController.startImageStream((image) => null);
-      cameraController.value =
-          cameraController.value.copyWith(isRecordingVideo: true);
-      expect(
-          cameraController.stopImageStream(),
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'A video recording is already started.',
-            'stopImageStream was called while a video is being recorded.',
-          )));
-    });
-
-    test('stopImageStream() throws $CameraException when not streaming images',
-        () async {
-      CameraController cameraController = CameraController(
-          CameraDescription(
-              name: 'cam',
-              lensDirection: CameraLensDirection.back,
-              sensorOrientation: 90),
-          ResolutionPreset.max);
-      await cameraController.initialize();
-
-      expect(
-          cameraController.stopImageStream(),
-          throwsA(isA<CameraException>().having(
-            (error) => error.description,
-            'No camera is streaming images',
-            'stopImageStream was called when no camera is streaming images.',
-          )));
-    });
-
-    test('stopImageStream() intended behaviour', () async {
-      MethodChannelMock methodChannelMock = MethodChannelMock(
-          channelName: 'plugins.flutter.io/camera',
-          methods: {'startImageStream': {}, 'stopImageStream': {}});
-      CameraController cameraController = CameraController(
-          CameraDescription(
-              name: 'cam',
-              lensDirection: CameraLensDirection.back,
-              sensorOrientation: 90),
-          ResolutionPreset.max);
-      await cameraController.initialize();
-      await cameraController.startImageStream((image) => null);
-      await cameraController.stopImageStream();
-      expect(methodChannelMock.log, <Matcher>[
-        isMethodCall('startImageStream', arguments: null),
-        isMethodCall('stopImageStream', arguments: null)
-      ]);
     });
 
     test('startVideoRecording() throws $CameraException when uninitialized',