+++ packages/camera/camera/test/camera_value_test.dart
@@ -5,6 +5,7 @@
 import 'dart:ui';
 
 import 'package:camera/camera.dart';
+import 'package:camera_platform_interface/camera_platform_interface.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter_test/flutter_test.dart';