+++ packages/camera/camera_platform_interface/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.3
+
+- Update Flutter SDK constraint.
+
+## 1.0.2
+
+- Added interface methods to support zoom features.
+
 ## 1.0.1
 
 - Added interface methods for setting flash mode.