+++ packages/connectivity/connectivity/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.0.3
+
+* Update Flutter SDK constraint.
+
 ## 2.0.2
 
 * Android: Fix IllegalArgumentException.