+++ packages/file_selector/file_selector/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.7.0+1
+
+* Update Flutter SDK constraint.
+
 ## 0.7.0
 
 * Initial Open Source release.