+++ packages/path_provider/path_provider_linux/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.1+3
+
+* Update Flutter SDK constraint.
+
 ## 0.1.1+2
 
 * Log errors in the example when calls to the `path_provider` fail.