+++ packages/url_launcher/url_launcher/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 5.7.13
+
+* Update Flutter SDK constraint.
+
+## 5.7.12
+
+* Updated code sample in `README.md`
+
 ## 5.7.11
 
 * Update integration test examples to use `testWidgets` instead of `test`.