+++ packages/video_player/video_player/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.2
+
+* Update Flutter SDK constraint.
+
 ## 1.0.1
 
 * Android: Dispose video players when app is closed.